requirements.txt
Flask==2.0.1
pymysql
flask_sqlalchemy
SQLAlchemy~=1.4.17
numpy~=1.20.1
pandas~=1.2.2
python-dateutil~=2.8.1